+++ src/bufferpool/FileBufferCache.h
@@ -7,7 +7,7 @@
 // Main interface combining cache and loader
 class FileBufferCache {
 public:
-    FileBufferCache(size_t maxCacheSize = 2 * 1024 * 1024 * 1024); // 2GB
+    FileBufferCache(uint64_t maxCacheSize = 2u * 1024 * 1024 * 1024); // 2GB
     ~FileBufferCache();
 
     // Fluent API entry point
